Support for new MasterCard BIN 2
"Mastercard® has received an additional range of 2-series numbers to add to the 5-series that we currently provide. The 2-series BINs operate in the same way as the 5-series and we have made changes to our systems to accommodate the new numbers. Readiness for the 2-series across our issuers, merchants and other partners is tracking on target dates listed below. Mastercard customer financial institutions can expect to be issued 2-series BINs starting in 2017."